Added Singleton Pattern and Comments to All My pages Plus sql and system action
Pages include: SqliteDataAccess, LogInPage, CreateCustomerAccountPage, CustomerHomePage, AccountInformationPage, CancelFlightPage, PaymentPage, BookFlightPage, and SystemAction

+++ ClassLibrary/FlightModel.cs
@@ -9,10 +9,8 @@ namespace ClassLibrary
 	public class FlightModel
 	{
 		// This class file is the FlightModel class. There are 
-		// the attributes associated with the Flight included, 
-		// the constructor to create an instance of the FlightModel,
-		// and methods to describe the actions the flight does.
-
+		// the attributes associated with the Flight included and  
+		// the constructor to create an instance of the FlightModel
 		// auto-implemented properties for trivial get and set
 		public int flightID { get; set; }
 		public int masterFlightID { get; set; }
@@ -24,13 +22,13 @@ namespace ClassLibrary
 		public DateTime departureDateTime { get; set; }
 		public TimeSpan duration { get; set; }
 		public string planeType { get; set; }
-		public int cost { get; set; }
+		public double cost { get; set; }
 		public int numOfPoints { get; set; }
 		public int numberOfVacantSeats { get; set; }
 		public double flightIncome { get; set; }
 
 		// customer constructor
-		public FlightModel(int fID, int mID, string origin, string oName, string destination, string dName, int dist, DateTime date, TimeSpan dur, string plane, int baseCost, int points, int seats, double income)
+		public FlightModel(int fID, int mID, string origin, string oName, string destination, string dName, int dist, DateTime date, TimeSpan dur, string plane, double baseCost, int points, int seats, double income)
 		{
 			flightID = fID;
 			masterFlightID = mID;
